deploy bundler helper
