+++ packages/boba/account-abstraction/package.json
@@ -65,6 +65,7 @@
     "hardhat-deploy": "^0.11.12",
     "hardhat-deploy-ethers": "^0.3.0-beta.11",
     "solidity-coverage": "^0.7.18",
+    "solidity-string-utils": "^0.0.8-0",
     "source-map-support": "^0.5.19",
     "table": "^6.8.0",
     "typescript": "^4.3.5"
